build: update SDK to 36 and improve payment logic
- Upgrade compileSdk and targetSdk to 36.
- Enhance previous operation detection with flexible field mapping.
- Update home navigation and standardize balance data types.

+++ shopik_flutter/lib/screens/statement_screen.dart
@@ -42,7 +42,7 @@ class _StatementScreenState extends State<StatementScreen> {
 
     final app = context.read<AppController>();
     setState(() {
-      _finalBalance = app.balance;
+      _finalBalance = app.balance.toDouble();
       _entries = [];
       _loading = false;
     });
